requirements.txt
requests==2.32.3
Mastodon.py==1.8.1
pydbus==0.6.0
